move intersect_curves from query to geom.py
